refactored BondListener to it's own class file, updated appropriate javadocs
