+++ app/src/main/java/com/idevicesinc/sweetblue/BleDeviceState.java
@@ -79,7 +79,7 @@ public enum BleDeviceState implements State
 	/**
 	 * Analogous to {@link BluetoothDevice#BOND_BONDING}. May not be relevant for your application if you don't use encrypted characteristics.
 	 * From this state, a device will either become {@link BleDeviceState#BONDED} (if successful) or {@link BleDeviceState#UNBONDED}.
-	 * If the latter, use {@link BleDevice.BondListener} to get further information on what happened.
+	 * If the latter, use {@link BondListener} to get further information on what happened.
 	 */
 	BONDING,
